packages/harness-core/pyproject.toml
[project]
name = "harness-core"
version = "0.1.0"
description = "General-purpose agentic ML framework"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "scikit-learn>=1.3",
    "scipy>=1.10",
    "omegaconf>=2.3",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
xgboost = ["xgboost>=1.7"]
catboost = ["catboost>=1.2"]
lightgbm = ["lightgbm>=4.0"]
neural = ["torch>=2.0", "pytorch-tabnet>=4.0"]
explore = ["optuna>=3.0"]
shap = ["shap>=0.42"]
viz = ["matplotlib>=3.7"]
quality = ["pandera>=0.17"]
notebook = ["nbformat>=5.9"]
drive = ["google-api-python-client>=2.100", "google-auth-oauthlib>=1.0"]
kaggle = ["kaggle>=1.6"]
all = [
    "harness-core[xgboost,catboost,lightgbm,neural,explore,shap,viz,quality,notebook,drive,kaggle]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harnessml"]
